requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication and security
PyJWT[crypto]==2.8.0
cryptography>=41.0.8
requests==2.31.0

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-core==2.4.1
google-auth==2.25.2

# Database and caching
redis==5.0.1
asyncio-redis==0.16.0

# Data processing
pandas>=2.1.4
numpy>=1.24.4,<2

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0